requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
supabase>=2.0.0
psycopg2-binary>=2.9.0
fastapi>=0.115.0
uvicorn>=0.32.0
gspread>=6.0.0
google-auth>=2.28.0
